refactor(services): PR-3.2 booru/ 子包搬迁 — 3 文件 + sys.modules 别名 shim
  booru_api.py   → booru/api.py
  booru_pool.py  → booru/pool.py
  downloader.py  → booru/downloader.py（项目级图片下载，实际只服务 booru）

回归：171 passed（含 booru_api / booru_pool / downloader / download_endpoints）
